perf(optim): Newton-Schulz in bf16 (modded-nanogpt style)
Port from origin/experimental: zeropower_via_newtonschulz5 uses bfloat16()
instead of float(). Cuts transient memory by ~2x in the orthogonalization
step and matches the nanoGPT speedrun path.

Peak memory: 12.18GB → 12.16GB (-0.02GB)
Profile: stable at 12.16GB for 5 steps (no leak)

Not-tested: convergence quality at bf16 vs fp32

+++ src/hagi/train/optim.py
@@ -15,7 +15,7 @@ def zeropower_via_newtonschulz5(G: torch.Tensor, steps: int = 5, eps: float = 1e
     """Approximate orthogonalization of a 2D matrix via quintic Newton-Schulz."""
     assert G.ndim == 2, "Muon orthogonalization expects a 2D matrix"
     a, b, c = (3.4445, -4.7750, 2.0315)
-    x = G.float()
+    x = G.bfloat16()
     transposed = G.size(0) > G.size(1)
     if transposed:
         x = x.T
